move functions to `aiida.tools.graph.deletions`
