+++ aiida/cmdline/commands/cmd_group.py
@@ -78,7 +78,7 @@ def group_remove_nodes(group, nodes, clear, force):
 def group_delete(group, clear, delete_nodes, dry_run, force, verbose, **traversal_rules):
     """Delete a group and (optionally) the nodes it contains."""
     from aiida.common.log import override_log_formatter_context
-    from aiida.manage.database.delete.nodes import delete_group_nodes, DELETE_LOGGER
+    from aiida.tools import delete_group_nodes, DELETE_LOGGER
     from aiida import orm
 
     if clear:
