+++ aiida/manage/database/delete/__init__.py
@@ -7,9 +7,3 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
-# pylint: disable=wildcard-import,undefined-variable,redefined-builtin,cyclic-import
-"""Functions for deleting entities from the database."""
-
-from .nodes import *
-
-__all__ = nodes.__all__
