+++ aiida/tools/graph/__init__.py
@@ -7,3 +7,8 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
+# pylint: disable=wildcard-import,undefined-variable
+"""Provides tools for traversing the provenance graph."""
+from .deletions import *
+
+__all__ = deletions.__all__
